Refactor: Organize files into a structured directory layout
This change introduces a more organized directory structure to the project by grouping files with similar functions into dedicated folders.

Key changes:
- Created a `src` directory to house all primary Python source code.
  - `src/core`: Contains core processing logic (e.g., `core_processor.py`).
  - `src/clients`: Contains API client implementations (e.g., `fanjiao_client.py`, `notion_client_cus.py`).
  - `src/server`: Contains server-related files (e.g., `webhook_server.py`, `api_server.py`).
  - `src/scripts`: Contains standalone scripts (e.g., `main.py`).
- Created a `tests` directory for test files.
- Moved relevant files into these new directories.
- Updated Python import statements across all affected files to reflect the new locations.
- Updated path references in `Dockerfile` and `compose.yml` to ensure the application can be built and run correctly with the new structure.

This reorganization aims to improve code maintainability and readability by establishing a clearer separation of concerns within the codebase.
