+++ compose.yml
@@ -16,4 +16,4 @@ services:
     volumes:
       - ./data:/app/data
     restart: unless-stopped
-    command: gunicorn --bind 0.0.0.0:5050 --workers 3 --timeout 120 webhook_server:app
+    command: gunicorn --bind 0.0.0.0:5050 --workers 3 --timeout 120 src.server.webhook_server:app
