+++ src/core/core_processor.py
@@ -8,9 +8,9 @@ from urllib.parse import urlparse
 
 from dotenv import load_dotenv
 
-from fanjiao_client import FanjiaoAPI, FanjiaoCVAPI
-from notion_client_cus import NotionClient
-from descrip_process import DescriptionProcessor
+from src.clients.fanjiao_client import FanjiaoAPI, FanjiaoCVAPI
+from src.clients.notion_client_cus import NotionClient
+from src.core.descrip_process import DescriptionProcessor
 
 # 仅本地开发时加载 .env 文件（Docker 环境会跳过）
 if os.getenv("ENV") != "production":
