+++ src/core/descrip_process.py
